More openvpnc code

+++ openvpnc.php
@@ -23,63 +23,66 @@
 	if(isset($_POST['button']))
 	{
 		$do = `sudo killall -TERM openvpn`;
-		sleep(2);
-
-		$file = escapeshellarg($_FILES['file']['tmp_name']);
-		$do = `sudo mv $file "/etc/openvpn/client/client1.ovpn"`;
-		$do = `sudo chown -R root: "/etc/openvpn/"`;
-
-		if(isset($_POST['passphrase']) || strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
+		if(isset($_FILES['file']['tmp_name']) && $_FILES['file']['tmp_name'] != "")
 		{
-			$passphrase = substr(escapeshellarg(trim($_POST['passphrase'])), 1, -1);
+			$file = escapeshellarg($_FILES['file']['tmp_name']);
+			$do = `sudo mv $file "/etc/openvpn/client/client1.ovpn"`;
+			$do = `sudo chown -R root: "/etc/openvpn/"`;
 
-			$ovpn = trim(`sudo cat "/etc/openvpn/client/client1.ovpn"`);
-			if(strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
+			if(isset($_POST['passphrase']) || strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
 			{
-				list($start, $rest) = explode("<key>", $ovpn, 2);
-				list($key, $rest) = explode("</key>", $rest, 2);
+				$passphrase = substr(escapeshellarg(trim($_POST['passphrase'])), 1, -1);
 
-				$start = trim($start);
-				$key = trim($key);
-				$rest = trim($rest);
-
-				$fn1 = tempnam("/tmp", "ovpn");
-				$fn2 = tempnam("/tmp", "ovpn");
-				$fp = fopen($fn1, "w");
-				fputs($fp, $key);
-				fclose($fp);
-
-				$cmd = "openssl rsa -in '$fn1' -out '$fn2' -passin 'pass:$passphrase'";
-				$do = `$cmd`;
-
-				$key = trim(`sudo cat '$fn2'`);
-				if($key != "")
+				$ovpn = trim(`sudo cat "/etc/openvpn/client/client1.ovpn"`);
+				if(strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
 				{
-					$ovpn = $start."\n<key>\n$key\n</key>\n$rest";
-					$do = `echo "$ovpn" | sudo tee "/etc/openvpn/client/client1.ovpn"`;
-				} else {
-					$errmsg = "Invalid passphrase supplied, please check your passphrase and retry uploading ovpn file.";
+					list($start, $rest) = explode("<key>", $ovpn, 2);
+					list($key, $rest) = explode("</key>", $rest, 2);
+
+					$start = trim($start);
+					$key = trim($key);
+					$rest = trim($rest);
+
+					$fn1 = tempnam("/tmp", "ovpn");
+					$fn2 = tempnam("/tmp", "ovpn");
+					$fp = fopen($fn1, "w");
+					fputs($fp, $key);
+					fclose($fp);
+
+					$cmd = "openssl rsa -in '$fn1' -out '$fn2' -passin 'pass:$passphrase'";
+					$do = `$cmd`;
+
+					$key = trim(`sudo cat '$fn2'`);
+					if($key != "")
+					{
+						$ovpn = $start."\n<key>\n$key\n</key>\n$rest";
+						$do = `echo "$ovpn" | sudo tee "/etc/openvpn/client/client1.ovpn"`;
+					} else {
+						$errmsg = "Invalid passphrase supplied, please check your passphrase and retry uploading ovpn file.";
+					}
+
+					unlink($fn1);
+					unlink($fn2);
 				}
-
-				unlink($fn1);
-				unlink($fn2);
 			}
-		}
 
-		if((!isset($_POST['passphrase']) || $_POST['passphrase'] == "") && strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
-		{
-			$errmsg = "ovpn file requires a passphrase, but you didn't supply one.";
+			if((!isset($_POST['passphrase']) || $_POST['passphrase'] == "") && strpos($ovpn, "-----BEGIN ENCRYPTED PRIVATE KEY-----") !== false)
+				$errmsg = "ovpn file requires a passphrase, but you didn't supply one.";
+		} else {
+			$errmsg = "No ovpn file was uploaded, can't continue";
 		}
 
 		if($errmsg == "")
 		{
 			if(file_exists("/etc/openvpn/client/client1.active"))
 			{
-				$do = `sudo /usr/sbin/openvpn --config "/etc/openvpn/client/client1.ovpn --daemon"`;
+				restartOVPN();
 				$okmsg = ".ovpn file was successfully imported and openvpn was successfully started";
 			} else {
 				$okmsg = ".ovpn file was successfully imported";
 			}
+		} else {
+			$do = `sudo rm -f "/etc/openvpn/client/client1.ovpn"`;
 		}
 	}
 
@@ -88,18 +91,26 @@
 		if(isset($_POST['enableCli']))
 		{
 			$do = `sudo touch "/etc/openvpn/client/client1.active"`;
-			$do = `sudo /usr/sbin/openvpn --config "/etc/openvpn/client/client1.ovpn --daemon"`;
+			restartOVPN();
+			$okmsg = "openvpn was successfully started";
 		} else {
 			$do = `sudo rm -f "/etc/openvpn/client/client1.active"`;
-			$do = `sudo killall -TERM openvpn`;
+			$do = `sudo killall openvpn`;
+
+			$okmsg = "openvpn was successfully stopped";
 		}
+	}
+
+	function restartOVPN()
+	{
+		if(intval(trim(`ps auxww|grep -v grep|grep openvpn|wc -l`)) > 0)
+			$do = `sudo killall openvpn`;
+		$do = `sudo /usr/sbin/openvpn --config "/etc/openvpn/client/client1.ovpn --daemon"`;
 	}
 
 	$enableCli = "no";
 	if(file_exists("/etc/openvpn/client/client1.active"))
-	{
 		$enableCli = "yes";
-	}
 
 	$page = 9;
 	$pageTitle = "OpenVPN Settings";
